Add documentation for `wallet_registerOnboarding` method
This method is used by our `@metamask/onboarding` library to facilitate
onboarding. It has been documented as an experimental API because it is
a public-facing method, though we have no immediate plans to
standardize this method, and we don't encourage its use directly.

A new section was created called "`wallet_` RPC methods" to
differentiate the two experimental RPC methods from the JavaScript
experimental provider API.