Refresh to MM Docs: glossary, intro material (#286)
This PR contains a significant revision to the Developer Documentation. There were a lot of links that were out of date. There are a lot of Dapp Developer resources available through the ConsenSys Academy that it made sense to provide links to. The Words Are Hard section now redirects to a ConsenSys glossary that includes the entries from the original glossary, as well as a lot more. Some descriptions were modified and broadened to improve clarity.

Co-authored-by: Erik Marks <25517051+rekmarks@users.noreply.github.com>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>