Update permission system documentation
The RPC API documentation about the permission system has been updated
to reflect the current terminology we use, and to reflect recent
changes.

The term "open methods" has been dropped in favor of "unrestricted
methods", which is more in-line with "restricted methods" and is what
we have been calling them in practice.

The wording was updated to no longer imply that a _domain_ is the
permission subject. Now it just says "caller" instead, which is just as
clear but leaves open the possibility that the caller is not a website.
This was done in preparation for Snaps, since snaps can also call RPC
methods.

The note on the implementation was removed, since the permission system
no longer uses `rpc-cap`. I removed it rather than linking the new
implementation because it's not really in a state where it would be
easy to understand or re-use.

The decryption methods were pulled into the "restricted methods"
section because they are restricted methods.

The "Other RPC Methods" header has been renamed to "Unrestricted
Methods". The old title was silly; this section was alongside the
header "Restricted Methods" so it always meant to refer to unrestricted
methods.