Recategorize encryption methods as unrestricted
The decryption-related methods have been moved back to the
"Unrestricted Methods" section because they have no corresponding
permissions. Instead the definition of "Unrestricted method" was
updated to no longer imply that no permissions were needed. Instead it
just means that there is no permission directly corresponding with that
RPC method.

In a later PR we will add more warnings about this constraint. We may
reconsider the terminology/categorization as well.