Update getting-started.md
Fleshing out the instructions for connecting to test networks and using Ganache. The instructions looked to be out of date refering to RPC etc (account is in top right, so made it clear with some screenshots) which is in the advanced section. Rather than refering to displaying of preconfigured test networks or just using the "Add Network" function in the network list.