GitHub Pages publish automation (#32)
* GitHub Pages publish automation

GitHub Actions have been added to automatically publish to GitHub
pages. The main branch is published to the path `/staging`, release
candidates are published to `/rc-[version]`, and published releases are
published to `/[version]` and `/[latest]`.

We can customize the paths/versioning later. We will also be adding
preview builds later.

* Simplify process by removing changelog and versioning

We don't need versioned docs at the moment, just a "staging" and
"production" site. The release process now consists just of merging
the `main` branch to `prod`.