Add preview builds (#44)
A preview build will now be triggered when a comment is left on a PR
starting with '@metamaskbot publish-preview'. The preview will be
published to the directory "preview/[commit hash]". The bot will
comment on the PR with a link to the preview once it has been
published.

Note that this only works on branches on this repository. For security
reasons, we don't allow previews on branches from forks.