Update numerous outdated examples in Anatomy of a Snap (#904)
* Better example of proposedName

proposedName in snap.manifest.json is supposed to be a human-readable name

also, Snap can be capitalized here

* Recommend ts instead of js

Our template now uses TypeScript by default so we should match that

* Do not call out that the SVG file is optional

It looks really bad when a Snap does not have an icon, so we should encourage developers to always have it

* More examples of proposedName

Better explain the difference between proposedName and package name. They are distinct.

* Replace deprecated wallet_enable example

wallet_requestSnaps is the right method to use here

* Simplify connect example

Also use a more standard variable name for the response from the snap invocation

* Align proposedName with manifest file example

* Double quotation marks for consistency

* Avoid confusing use of "it" with multiple subjects

* Update config example to ts

Also remove the browserify instructions (to be rewritten for webpack later) and add a note about bundling with the template.